Refactor project structure: move components to features directory
