--- src/components/Navigation/DarkModeToggle.tsx
@@ -1,20 +1,0 @@
-import React from 'react';
-import { Moon, Sun } from 'lucide-react';
-
-interface DarkModeToggleProps {
-  isDarkMode: boolean;
-  setIsDarkMode: (isDarkMode: boolean) => void;
-}
-
-const DarkModeToggle: React.FC<DarkModeToggleProps> = ({ isDarkMode, setIsDarkMode }) => {
-  return (
-    <button
-      onClick={() => setIsDarkMode(!isDarkMode)}     
-      className="p-2 text-gray-600 dark:text-gray-400 hover:text-blue-600 dark:hover:text-blue-400 hover:bg-blue-50 dark:hover:bg-blue-900/30 rounded-lg transition-all duration-200"
-    >
-      {isDarkMode ? <Sun className="h-5 w-5" /> : <Moon className="h-5 w-5" />}
-    </button>
-  );
-};
-
-export default DarkModeToggle;
