+++ src/features/properties/AdditionalInfoSection.tsx
