+++ src/features/properties/AmenitiesSection.tsx
