+++ src/features/properties/DescriptionSection.tsx
