+++ src/features/properties/ImageGallery.tsx
