+++ src/features/properties/ImageUploadSection.tsx
