+++ src/features/properties/OwnerDetails.tsx
