+++ src/features/properties/PropertyCard.tsx
