+++ src/features/properties/PropertyDetails.tsx
