+++ src/features/properties/PropertyDetailsSection.tsx
