+++ src/features/properties/PropertyListItem.tsx
